13.03

+++ Blog.Api/Managers/BlogManager/BlogManager.cs
@@ -1,6 +1,7 @@
 ﻿using Blog.Api.Context;
 using Blog.Api.Entities;
 using Blog.Api.Models.BlogModels;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.EntityFrameworkCore;
 
 namespace Blog.Api.Managers.BlogManager;
@@ -16,15 +17,21 @@ public class BlogManager
 
     public async Task<List<BlogModel>> GetBlogs()
     {
-        var blogs = await _dbContext.Blogs.Include(b => b.Posts).ToListAsync();
-
-        return ParseList(blogs);
+        try
+        {
+            var blogs = await _dbContext.Blogs.Include(b => b.Posts).ToListAsync();
+            return ParseList(blogs);
+        }
+        catch (Exception ex)
+        {
+            return new List<BlogModel>(); 
+        }
     }
 
     public async Task<BlogModel> GetBlogById(Guid id)
     {
-        var blog = IsExist(id);
-
+        var blog = _dbContext.Blogs.FirstOrDefault(b => b.Id == id);
+        if (blog == null) throw new Exception("Not found");
         return ParseToBlogModel(blog);
     }
 
@@ -58,7 +65,9 @@ public class BlogManager
     }
     public async Task<BlogModel> UpdateBlog(Guid blogId, CreateBlogModel model)
     {
-        var blog = IsExist(blogId);
+        var blog = _dbContext.Blogs.FirstOrDefault(b => b.Id == blogId);
+        if (blog == null) throw new Exception("Not found");
+
         blog.Name = model.Name;
         blog.Description = model.Description;
         await _dbContext.SaveChangesAsync();
@@ -67,10 +76,16 @@ public class BlogManager
 
     public async Task<string> DeleteBlog(Guid blogId)
     {
-        var blog = IsExist(blogId);
+        var blog = await _dbContext.Blogs.FindAsync(blogId);
+
+        if (blog == null)
+        {
+            throw new Exception("Blog not found."); 
+        }
+
         _dbContext.Blogs.Remove(blog);
         await _dbContext.SaveChangesAsync();
-        return "Done :)";
+        return "No content";
     }
 
     private BlogModel ParseToBlogModel(Entities.Blog blog)
@@ -91,7 +106,12 @@ public class BlogManager
     private string GetUserName(Guid userId)
     {
         var user = _dbContext.Users.Find(userId);
-        return user!.Username;
+
+        if (user == null)
+        {
+            throw new Exception("User not found."); 
+        }
+        return user.Username;
     }
 
     private List<BlogModel> ParseList(List<Blog.Api.Entities.Blog> blogs)
@@ -104,10 +124,5 @@ public class BlogManager
         return blogModels;
     }
 
-    private Entities.Blog IsExist(Guid blogId)
-    {
-        var blog = _dbContext.Blogs.FirstOrDefault(b => b.Id == blogId);
-        if (blog == null) throw new Exception("Not found");
-        return blog;
-    }
+    
 }
